Track redirects via onBeforeNavigate
Normally if you receive a remote URL, the transition type will be
"link", but the latest visit for that page will have a referringVisitId
of -1, because you didn't actually click on a link to make that
navigation happen.  One exception to this is redirects - if I send, say,
https://ddg.gg (which redirects to https://duckduckgo.com), onCommitted
only picks up the redirect destination, and the visit corresponding to
the original URL is considered to be the referring visit.  So if we
detect a redirect, we need to find out the visit that brought us here.

To do that, we can subscribe to the onBeforeNavigate event, which
gets the original URL.  We can track the previous URL in a table (keyed
by tab, since redirects *should* happen within the same tab) and use
that to look up the original visit to see if *it* has no
referringVisitId